[quartermaster] refactor: Split large files into partial classes (#769)
Split 4 large files exceeding 500-line threshold into partial classes for improved maintainability and AI readability:

- MdlAsciiReader: 1083 → 7 files (~200 lines each)
- MdlBinaryReader: 1071 → 8 files (~150 lines each)
- AppearancePanel: 755 → 3 files (~250 lines each)
- MainWindow.Inventory: 614 → 3 files (~200 lines each)

All tests pass (322/322).
